requirements.txt
opencv-python==4.6.0.66
Pillow==10.1.0
pyvirtualcam==0.11.0